refactor(entity): rename `appointments.ts` to  `appointment.ts`
