+++ src/domain/schemas/appointment.ts
